Add handling of redirects
The ReverseProxy does not directly handle redirects and, instead, sends the
3xx response back to the caller. If the Location header provided by the redirect
is absolute (http://localhost/path/to/api) instead of relative (/path/to/api) it
results in clients making a second request to the remote system instead of making
another requests through the proxy. This addition ensures that all Location
headers returned by the proxy contain relative URLs so subsequent requests are
always targeted to the proxy.