Move hostRewrite decorator to beginning of chain
This was previously handled by wrapping the base transport directly
and allowing all other features to layer on top of this core function.
This was done for two reasons: 1) the rewrite data were most available
in the same portion of the startup process as creating the base transports
and 2) it ensured that the host rewrite could not be altered by other
middleware.

Unfortunately, there are some cases where code/tools need access to the
rewritten host in the request object when they execute. This came up during
an addition to the serverfull-gateway project where AWS requires request
sigining that includes the target host. It is likely that any other API that
requires request signing will need the same information.

This change results in the host rewrite happening before all other decorators
so that they have access to the rewritten host name and URL data. This does enable
future middleware to interfer in the process and rewrite the host yet again.