Allow unmatched routes for the default backend
I'm attempting to introduce transportd for a series of services that
either have no OpenAPI specification or have one that is not guaranteed
to be correct. I'm looking to add an option that provides a bit of
forgiveness on an opt-in basis so that my teams can get started with the
proxy without needing their contracts to be perfect yet. I plan to
introduce the proxy with additional middleware that report on route
mismatches rather than denying them outright so folks have time to
update their specs.

This patch introduces an option to allow requests that do not match any
given OpenAPI path to still be proxied to a backend. This works by
establishing a canonical name for a default backend. Realisically,
there's no way to route unmatched paths to different backends as that is
among the core purposes of the route mapping to begin with. Establishing
that one and only one backend may receive unmatched routes resolves the
conflict. The next version of the OpenAPI spec _should_ include the
ability to use wildcards in routes in a way that would replace this
feature over time. There is no timeline for when that next version will
be available.

This patch ensures that anyone currently using the name "default", the
canon name for the default backend, is unaffected. The new behavior must
be opted into by adding a new section called "allowUnknown" which
contains a full suite of route settings that will be applied for all
unknown routes. This makes it possible to log, stat, etc., the unknown
routes before proxying them to the backend. The only difference between
the "allowUnknown" section and the "x-transportd" section typically
found under a route is that "allowUnknown" does not require a "backend"
key and any key of that name given is overwritten with "default". I did
this to avoid requiring the user to denote an obvious relationship.