Update to latest kin openapi
The openapi library maintainer added a second option for a router that
is build on gorilla/mux. When they did this they re-arranged some of the
locations of types to offer a better abstraction over the router.
Additionally, they removed the term Swagger from all methods and types.

Generally, everything is still the same shape as before but the names
changed. The only part of this project that exposes the openapi library
types directly are the context methods in the clienttransport.go file.
Because these are exposed directly this is technically a breaking change
even though the new exposed types are identical to the old ones. The
only place within the project that these are used is the validation
middleware which worked without modifications.